fix: break pushNotifications/LocaleContext/useStores require cycle
