+++ src/services/pushNotifications.ts
@@ -1,12 +1,9 @@
-import { useEffect } from "react";
 import { Alert } from "react-native";
 import * as Device from "expo-device";
 import * as Notifications from "expo-notifications";
 import { Colors } from "@common-ui/constants/colors";
-import { useRouter } from "expo-router";
 import { isAndroid, isWeb } from "@common-ui/utils/responsive";
 import { openAppSettings } from "@utils/navigation";
-import { useLocale } from "@common-ui/contexts/LocaleContext";
 
 // This is for foreground notifications
 Notifications.setNotificationHandler({
@@ -70,46 +67,4 @@ export async function registerForPushNotificationsAsync(
   }
 
   return token;
-}
-
-export function useRegisterPushNotificationsListener(requestPermissions: boolean) {
-  const router = useRouter();
-  const { t } = useLocale();
-
-  useEffect(() => {
-    registerForPushNotificationsAsync(requestPermissions, t);
-    // Clear badge count on app open
-    Notifications.setBadgeCountAsync(0);
-  }, []);
-
-  useEffect(() => {
-    let isMounted = true;
-
-    if (isWeb) {
-      return () => {};
-    }
-
-    function redirect(notification: Notifications.Notification) {
-      const url = notification.request.content.data?.url || notification.request.content.data?.path;
-      if (url) {
-        router.push(url as any);
-      }
-    }
-
-    Notifications.getLastNotificationResponseAsync().then((response) => {
-      if (!isMounted || !response?.notification) {
-        return;
-      }
-      redirect(response?.notification);
-    });
-
-    const subscription = Notifications.addNotificationResponseReceivedListener((response) => {
-      redirect(response.notification);
-    });
-
-    return () => {
-      isMounted = false;
-      subscription.remove();
-    };
-  }, []);
 }
